airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astro-sdk-python[snowflake] >= 1.1.0
astro-sdk-python